requirements.txt
mcp>=1.0.0
pymysql>=1.1.0
cryptography>=41.0.0
python-dotenv>=1.0.0
